python-transcription-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==11.0.3
speech-recognition==3.10.0
pydub==0.25.1
openai-whisper==20231117
torch==2.1.0
numpy==1.24.3
python-multipart==0.0.6
redis==5.0.1
pymongo==4.6.0
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4